Add explorable explanation heading.
